# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Resume {
  id: number;
  name: string;
  content: string;
  fileName: string;
  createdAt: string;
}

export interface CreateResumeBody {
  name: string;
  content: string;
  fileName: string;
}

export interface AnalyzeResumeBody {
  resumeId: number;
  jobDescription: string;
}

export interface AnalysisResult {
  /** Match percentage 0-100 */
  matchScore: number;
  matchedSkills: string[];
  missingSkills: string[];
  suggestions: string[];
  summary: string;
  analysisId: number;
}

export interface AnalysisRecord {
  id: number;
  resumeId: number;
  jobDescription: string;
  matchScore: number;
  matchedSkills: string[];
  missingSkills: string[];
  suggestions: string[];
  summary: string;
  createdAt: string;
}