# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all resumes
 */
export const ListResumesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  content: zod.string(),
  fileName: zod.string(),
  createdAt: zod.date(),
});
export const ListResumesResponse = zod.array(ListResumesResponseItem);

/**
 * @summary Upload a resume
 */
export const CreateResumeBody = zod.object({
  name: zod.string(),
  content: zod.string(),
  fileName: zod.string(),
});

/**
 * @summary Get a resume by ID
 */
export const GetResumeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetResumeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  content: zod.string(),
  fileName: zod.string(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a resume
 */
export const DeleteResumeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Analyze resume against a job description
 */
export const AnalyzeResumeBody = zod.object({
  resumeId: zod.number(),
  jobDescription: zod.string(),
});

export const AnalyzeResumeResponse = zod.object({
  matchScore: zod.number().describe("Match percentage 0-100"),
  matchedSkills: zod.array(zod.string()),
  missingSkills: zod.array(zod.string()),
  suggestions: zod.array(zod.string()),
  summary: zod.string(),
  analysisId: zod.number(),
});

/**
 * @summary Get analysis history for a resume
 */
export const GetAnalysisHistoryParams = zod.object({
  resumeId: zod.coerce.number(),
});

export const GetAnalysisHistoryResponseItem = zod.object({
  id: zod.number(),
  resumeId: zod.number(),
  jobDescription: zod.string(),
  matchScore: zod.number(),
  matchedSkills: zod.array(zod.string()),
  missingSkills: zod.array(zod.string()),
  suggestions: zod.array(zod.string()),
  summary: zod.string(),
  createdAt: zod.date(),
});
export const GetAnalysisHistoryResponse = zod.array(
  GetAnalysisHistoryResponseItem,
);