# lib/api-zod/src/generated/types/analyzeResumeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeResumeBody {
  resumeId: number;
  jobDescription: string;
}