# lib/api-zod/src/generated/types/createResumeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateResumeBody {
  name: string;
  content: string;
  fileName: string;
}